TST: refactor tests
re-work openstack api security group tests to work with new functions
